+++ packages/public.config/src/types/index.d.ts
@@ -5,8 +5,11 @@ declare global {
     BACKEND?: {
       baseURL?: string;
     };
+    GIT_AUTHOR_DATE?: string;
+    GIT_VERSION?: string;
     KEYCLOAK: Parameters<typeof Keycloak>[0];
     KEYCLOAK_MOCK_USER?: MockUser;
     LOCALES?: string[];
+    PROD: boolean;
   } | void;
 }