+++ packages/react-app-locale-utils/src/types/index.d.ts
@@ -1,0 +1,7 @@
+declare module '@visma/vite-plugin-date-fns-locale/dynamic-import-date-fns-locales' {
+  interface T {
+    [locale: string]: () => Promise<{ default: import('date-fns').Locale }>;
+  }
+  const dateFnsLocaleModules: T;
+  export default dateFnsLocaleModules;
+}