+++ packages/react-app-super-template/src/main.tsx
@@ -1,1 +1,3 @@
+import './patchGlobals.js';
+
 export { default } from './AppProvider.js';