+++ packages/react-app-super-template/src/patchGlobals.ts
@@ -1,0 +1,13 @@
+import { Buffer } from 'buffer';
+import EventsEventEmitter from 'events';
+import process from 'process';
+
+if (ENV?.PROD) {
+  (globalThis as any).Buffer = Buffer;
+  globalThis.process = process;
+  globalThis.EventEmitter = EventsEventEmitter;
+}
+
+declare global {
+  var EventEmitter: typeof EventsEventEmitter;
+}