+++ packages/vite-plugin-super-template/src/defaultExport.ts
@@ -1,0 +1,4 @@
+export default function defaultExport<T>(module: T) {
+  const cjsModule: { default: T } = module as any;
+  return typeof cjsModule?.default === 'undefined' ? module : cjsModule.default;
+}