+++ packages/react-intl-bundled-messages/CHANGELOG.md
@@ -1,5 +1,10 @@
+# 1.1.0
+
+- Add `@visma/vite-plugin-react-intl-bundled-messages` (internal)
+
 # 1.0.0
 
+- **Breaking:** Switch to Vite. Remove support of `useMessages` in CRACO.
 - Upgrade to `@visma/react-app-locale-utils@^1.0.0`
 
 # 0.1.5